Backported support for all integral types to legacy frameworks via source generation
Introduces an incremental source generator for SqidsEncoder that adds overloads supporting all integral types on legacy frameworks.

Updates README with usage details, adds new tests for legacy encoding/decoding, and wires up the code generation project in the solution and main project.

This enables drop-in support for byte, sbyte, short, ushort, int, uint, long, and ulong encoding/decoding in .NET Standard 2.0 (.NET Framework 4.7.2+) and .NET 6 environments.

Extreme care was taken to avoid structural and code changes to the SqidsEncoder class (even comments were kept where they were, as a diff will show) to make it easy to keep it up to date with the main Sqids project if a PR is not accepted.