Add Sqids.Benchmarks project with initial benchmarks
Introduces a new Sqids.Benchmarks project to the solution for performance testing. Includes a custom BenchmarkDotNet config and custom project variables that allow referencing the old and new versions of the package conditionally, and also allow to create conditional code based on these variables to properly compare performance across versions of the package.